write SSM parameter name to config file
the round trips to IAM to fetch the role and policy to determine the
name of the SSM parameter that holds the lambda config can drive the
coldstart initialisation time for the lambda beyond 5 seconds, which
then causes cloudfront to throw a 503 error. This change writes the name
of the SSM parameter to a local config file which is deployed with the
lambda code.  When invoked by cloudfront the lambda reads the config
file to determine the name of the SSM parameter to fetch.  This reduces
the number of round trips to IAM, which speeds up cold starts and
reduces the 503 error rate.