Add option to package config in lambda (#143)
add options to include config in lambda package

IAM introspection and config fetch from SSM in us-east-1 can be slow,
which causes sporadic 503 errors in regions far from us-east-1.

This change adds an input lambda_config_mode which causes the json
config to be written to a file packaged/deployed with the lambda code -
see the comments in the README for further rationale.

default behaviour is to continue to use IAM introspection, so this
should be a non-breaking change.